Resize image and return a token in response
