+++ api/__init__.py
@@ -1,7 +1,7 @@
 from flask_restful import Api
 
 from app import flaskAppInstance
-from .Image import Image
+from .ResizeImage import ResizeImage
 
 restServer = Api(flaskAppInstance)
-restServer.add_resource(Image, "/api/v1/resize")
+restServer.add_resource(ResizeImage, "/api/v1/resize")
